Remove key/value store
Since we were only using the key/value store to hold the last check
timestamp, we removed it. It was only needed for the local event store,
so we have made the last check timestamp an atom that is a dependency
for the local event store instead. This also addresses #11