Encrypts data at rest
Closes #21

- Adds encryption to SQL object store so that the entire object is
  encrypted
- Adds encryption to SQL event store so that the event data is encrpyted

This does not affect the mongodb or in-memory stores. Ideally, we should
make it configurable whether to encrypt data or not. Also, the fact that
we are constantly updating (potentially) large encrypted blobs whenever
we emit an event to a job highlights the fact that we may want to be a
bit smarter about what we store and even normalize in places so that we
can perform more surgical updates and leave most data unencrypted. Jobs
should probably be stored on their own rather than as "blobs".